start working on SOL::C::* objects
SOL::C::* objects represent exactly the s_* structures in the Neverball
code. That is, they are not converted into a form that would be easier
to work with in Perl. They are essentially just some data storing hashrefs
that know how they are represented in a SOL file.

Only little changes are allowed, like representing a flag field as strings,
or replacing -1 with undef where an index is expected. Conversion of
coordinates is not done, for example.